Fixes Production APK builds by using reflection on MockRepositoryUtility
The absence of this class on the main source set causes crashes
on the apk signature as the import statement points to a
non-existing class. However, it is desirable to keep it separated
from the main source as its only purpose is to help testing, and as
such should never be included in a production release.
Reflection allows us to attempt to load it on runtime and catches
its absence to fallback to a production-ready behavior.