Fixes delay in closing of ArchivedAccountsDialogFragment
Note: Dialog is now closed immediately after user clicks "Yes"
on confirmation dialog for account deletion. While this might
lead to some unexpected behavior, worst case scenario is dialog
closing without the account being deleted (due to some database error)
but I don't consider that to be that bad of a bug compared to the
visible delay that was being experienced before.