Turns MonthlyBalance.date into SQL index
By creating a test SQLite database (on PC) and running
MonthlyBalance related queries on it, it turns out that,
despite balance_date being part of a primary key, it is not
used automatically as an index. As all queries on this
entity rely on some filtering based on balance_date,
it's important to avoid sequential searches on this table.