Delegates group responsibility from TransactionsListViewModel to TransactionsFragment
Previous implementation ended up looping through transactions set twice.
One to generate the map, and another one to display it. It was also
kinda redundant since repository already returns a sorted list.

New implementation, while a bit harder to read, avoids that and opens
the door for alternative sorting methods (that are not linked to the
transaction's date)